[FIX] Properly compare version numbers
To put you in context, I'm preparing a release for Odoo 13.0, but currently the closest thing we have is saas-12.5.

[In this version, `odoo.release.version_info[0]` is `"saas~12.5"`][1], which [produces errors][2] when using click-odoo because it's comparing an int to a str:

```
2019-09-11 12:43:49,398 1 ERROR ? click_odoo.env_options: exception
Traceback (most recent call last):
  File "/usr/local/lib/python3.7/site-packages/click_odoo/env_options.py", line 178, in _invoke
    self._configure_odoo(ctx)
  File "/usr/local/lib/python3.7/site-packages/click_odoo/env_options.py", line 155, in _configure_odoo
    self._fix_odoo_logging()
  File "/usr/local/lib/python3.7/site-packages/click_odoo/env_options.py", line 115, in _fix_odoo_logging
    if odoo.release.version_info[0] < 9:
TypeError: '<' not supported between instances of 'str' and 'int'
Error: '<' not supported between instances of 'str' and 'int'
```

I realized that not only click-odoo uses this system, but click-odoo-contrib also.

Reacting to Odoo version is a very important part of click-odoo scripting, so it seems better IMHO to supply a `click_odoo.parse_version()` function that, for now, just inherits Odoo's own method. In the future, if this method becomes inconsistent in Odoo, click-odoo would be the abstraction layer to preserve safety in click-odoo scripts. For now, it's consistent across all supported versions, so importing it should be enough.

[1]: https://github.com/odoo/odoo/blob/bc407fd18e68766bd49c772de7403d3d96dc3d47/odoo/release.py#L15
[2]: https://travis-ci.org/Tecnativa/doodba/jobs/583633629#L8708-L8717